+++ crates/generic_log_worker/src/log_ops.rs
@@ -39,9 +39,9 @@ use std::{
 };
 use thiserror::Error;
 use tlog_tiles::{
-    ConsistencyProof, Hash, HashReader, InclusionProof, LogEntry, PendingLogEntry, Tile,
-    TileHashReader, TileIterator, TileReader, TlogError, TlogTile, TreeWithTimestamp,
-    UnixTimestamp, HASH_SIZE,
+    ConsistencyProof, Hash, HashReader, InclusionProof, LogEntry, PendingLogEntry,
+    PreloadedTlogTileReader, TileHashReader, TileIterator, TlogError, TlogTile, TlogTileRecorder,
+    TreeWithTimestamp, UnixTimestamp, HASH_SIZE,
 };
 use tokio::sync::watch::{channel, Receiver, Sender};
 use worker::Error as WorkerError;
@@ -228,100 +228,6 @@ pub(crate) struct SequenceState {
 struct TileWithBytes {
     tile: TlogTile,
     b: Vec<u8>,
-}
-
-/// A fake `TileReader` that just records the tiles that are requested, but
-/// doesn't actually read the tile data.
-#[derive(Default)]
-struct TlogTileRecorder(RefCell<Vec<TlogTile>>);
-
-impl TileReader for TlogTileRecorder {
-    fn height(&self) -> u8 {
-        TlogTile::HEIGHT
-    }
-
-    /// Records the requested tree tiles without actually reading them, and
-    /// always returns an error.
-    ///
-    /// # Errors
-    ///
-    /// Will return a `TlogError::InvalidInput` if any of the requested tiles is
-    /// not a valid `TlogTile`. Otherwise, always returns a
-    /// `TlogError::RecordedTilesOnly` in compliance with the `TileReader`
-    /// contract.
-    fn read_tiles(&self, tiles: &[Tile]) -> Result<Vec<Vec<u8>>, TlogError> {
-        // Record the tiles we're meant to read. Convert them to tlog tiles,
-        // ensuring their height is always 8
-        *self.0.borrow_mut() = tiles
-            .iter()
-            .map(|t| {
-                if t.height() == TlogTile::HEIGHT {
-                    Ok(TlogTile::new(t.level(), t.level_index(), t.width(), None))
-                } else {
-                    Err(TlogError::InvalidInput(
-                        "TlogTileRecorder cannot read tiles of height not equal to 8".to_string(),
-                    ))
-                }
-            })
-            .collect::<Result<Vec<_>, TlogError>>()?;
-
-        // Return an error since we did not actually read the tiles.
-        Err(TlogError::RecordedTilesOnly)
-    }
-
-    // Do nothing; we only use this struct to record tiles.
-    fn save_tiles(&self, _tiles: &[Tile], _data: &[Vec<u8>]) {}
-}
-
-/// A thin wrapper around a map of tlog tile ⇒ bytestring. Implements
-/// `TileReader` so we can use it within `TileHashReader::read_tiles` to read
-/// and verify tiles.
-struct PreloadedTlogTileReader(HashMap<TlogTile, Vec<u8>>);
-
-impl TileReader for PreloadedTlogTileReader {
-    fn height(&self) -> u8 {
-        TlogTile::HEIGHT
-    }
-
-    /// Converts the given tiles into tlog tiles, then reads them from the
-    /// internal hashmap.
-    ///
-    /// # Errors
-    /// Errors if any of given tiles hash height != 8, or is a data tile. Also
-    /// errors if the hash map is missing tiles from the input here.
-    fn read_tiles(&self, tiles: &[Tile]) -> Result<Vec<Vec<u8>>, TlogError> {
-        let mut buf = Vec::with_capacity(HASH_SIZE * TlogTile::FULL_WIDTH as usize);
-
-        for tile in tiles {
-            // Convert the tile to a tlog-tile, ie one where height=8 and data=false
-            if tile.height() != TlogTile::HEIGHT {
-                return Err(TlogError::InvalidInput(
-                    "PreloadedTlogTileReader cannot read tiles of height not equal to 8"
-                        .to_string(),
-                ));
-            }
-            if tile.is_data() {
-                return Err(TlogError::InvalidInput(
-                    "PreloadedTlogTileReader cannot read data tiles".to_string(),
-                ));
-            }
-            let tlog_tile = TlogTile::new(tile.level(), tile.level_index(), tile.width(), None);
-
-            // Record the tile's contents
-            let Some(contents) = self.0.get(&tlog_tile) else {
-                return Err(TlogError::InvalidInput(format!(
-                    "PreloadedTlogTileReader cannot find {}",
-                    tlog_tile.path()
-                )));
-            };
-            buf.push(contents.clone());
-        }
-
-        Ok(buf)
-    }
-
-    /// Do nothing; we only use this struct to read tiles.
-    fn save_tiles(&self, _tiles: &[Tile], _data: &[Vec<u8>]) {}
 }
 
 /// An error that can occur when creating a log.
